adjust outer join behavior (types and right outer join bug)
outer joins need to return nullable tables as they may introduce missing
data. similar_nullable on DataTables has been removed (unused) and
replaced with a similar_nullable that works on
NullableCategoricalArrays, and this change is made to support the new
changes to join. The 3 outer joins share a function with inner joins,
and this shared function (compose_joined_table) now performs a check to
see if the join type is :inner, and if so, it will return the same
column type as the parent table rather than promoting to a nullable
column. A bug was found in right-outer join behavior where the values
unique to the right table were added to the table in the incorrect
locations, overwriting data and leaving nulls where they shouldn't be.
This bug, due to incorrect values in rightonly_ixs.join, was fixed by
filling the last n-rows of the datatable where n =
length(rightonly_ixs.join). Tests were checked for accuracy against
pandas.