Consolidate DataTable constructors and remove autopromotion
Consolidating the constructors minimized the number of places where
auto promotion could take place. The new constructor recycles scalars
such that if DataTable is created with a mix of scalars and vectors
the scalars will be recycled to the same length as the vectors. Fixes an
outstanding bug where scalar recycling only worked if the scalar
assignments came after the vector assignments of the desired length, see
https://github.com/JuliaStats/DataFrames.jl/pull/882. Tests that
used to assume NullableArray promotion now explicitly use NullableArrays
and new constructor tests have been added to test changes.